deleted the deleteLunch method in LunchAdminBean and using the one in LunchAPI instead

+++ src/main/java/se/miun/dt170g/projektdt170g/API/LunchAPI.java
@@ -55,7 +55,7 @@ public class LunchAPI implements Serializable{
 
     @DELETE
     @Path("/{id}")
-    public Response deleteLunch(@PathParam("id") Long id) {
+    public Response deleteLunch(@PathParam("id") int id) {
         try {
             LunchMenuEntity lunchMenu = entityManager.find(LunchMenuEntity.class, id);
             if (lunchMenu != null) {
